Final RTAI 4.9.9 release
This will be the last release of the RTAI 4 branch.
The 3.10 and 3.18 kernels have reached EOL, and the
3.16.53 kernel has introduced assembly changes that are
too complicated for me to work around.

All future development will now be focused on RTAI 5,
with performance improvements and clean-ups as always.

Special thanks to jepler and seb_kuzminsky!

Nothing left to do.

Signed-off-by: Alec Ari <neotheuser@ymail.com>