rtai_math.ko and musl: Switch from gnu89/C99 to GNU11
The RTAI math kernel module (rtai_math.ko) was originally
built with gnu89, with musl itself being built with C99.
This is because musl is built outside of Kbuild
(therefore not a kernel module) and then is linked to
rtai_math.ko as the final step.

musl's libm libary provides the nitty gritty of math functions
which are used and exported by rtai_math.ko so it can be used
by external projects such as LinuxCNC.

This commit forces both the RTAI math module (rtai_math.ko)
and musl itself to pure GNU11.

With this, no modification to LinuxCNC is required. LinuxCNC
uses `rtai-config --module-cflags` to capture the appropriate
compiler flags to pass to it's Kbuild system (what builds the
LinuxCNC kernel modules.) The `--module-cflags` option
represents RTAI_FP_CFLAGS.

Signed-off-by: Alec Ari <neotheuser@ymail.com>